reduced, renamed and simplified android before adding Flutter
