+++ add_to_app/android_view/app/src/main/java/dev/flutter/example/androidView/ui/docs/DocsFragment.kt
@@ -1,0 +1,34 @@
+package dev.flutter.example.androidView.ui.docs
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import dev.flutter.example.androidView.databinding.FragmentDocsBinding
+
+class DocsFragment : Fragment() {
+    private var _binding: FragmentDocsBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentDocsBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
