+++ add_to_app/android_view/app/src/main/java/dev/flutter/example/androidView/ui/feed/FeedFragment.kt
@@ -1,0 +1,35 @@
+package dev.flutter.example.androidView.ui.feed
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import dev.flutter.example.androidView.databinding.FragmentFeedBinding
+
+class FeedFragment : Fragment() {
+    // TODO: add toggle for single vs multi-Flutter cells.
+    private var _binding: FragmentFeedBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentFeedBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
