--- add_to_app/android_view/app/src/main/java/dev/flutter/example/androidView/ui/notifications/NotificationsViewModel.kt
@@ -1,13 +1,0 @@
-package dev.flutter.example.androidView.ui.notifications
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class NotificationsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
-    }
-    val text: LiveData<String> = _text
-}
