+++ add_to_app/android_view/app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,18 +2,13 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
-
-    <item
-        android:id="@+id/navigation_dashboard"
+        android:id="@+id/navigation_feed"
         android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:title="@string/title_feed" />
 
     <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:id="@+id/navigation_docs"
+        android:icon="@drawable/ic_home_black_24dp"
+        android:title="@string/title_docs" /> <!-- TODO: Find a better icon-->
 
 </menu>
